+++ src/deluge/dsp/master_compressor/master_compressor.cpp
@@ -78,71 +78,3 @@ void MasterCompressor::setup(int32_t attack, int32_t release, int32_t threshold,
 	setMakeup((float)makeup / 100.0);
 	wet = (float)makeup / 100.0;
 }
-namespace chunkware_simple {
-//-------------------------------------------------------------
-// envelope detector
-//-------------------------------------------------------------
-EnvelopeDetector::EnvelopeDetector(float timeConstant, float sampleRate) {
-	assert(sampleRate > 0.0);
-	assert(timeConstant > 0.0);
-	sampleRate_ = sampleRate;
-	timeConstant_ = timeConstant;
-	setCoef();
-}
-//-------------------------------------------------------------
-void EnvelopeDetector::setTc(float timeConstant) {
-	assert(timeConstant > 0.0);
-	timeConstant_ = timeConstant;
-	setCoef();
-}
-//-------------------------------------------------------------
-void EnvelopeDetector::setSampleRate(float sampleRate) {
-	assert(sampleRate > 0.0);
-	sampleRate_ = sampleRate;
-	setCoef();
-}
-//-------------------------------------------------------------
-void EnvelopeDetector::setCoef(void) {
-	nSamplesInverse_ = exp(-1000.0 / (timeConstant_ * sampleRate_));
-}
-
-//-------------------------------------------------------------
-// attack/release envelope
-//-------------------------------------------------------------
-AttRelEnvelope::AttRelEnvelope(float att_ms, float rel_ms, float sampleRate)
-    : attackEnvelope_(att_ms, sampleRate), releaseEnvelope_(rel_ms, sampleRate) {
-}
-//-------------------------------------------------------------
-void AttRelEnvelope::setAttack(float ms) {
-	attackEnvelope_.setTc(ms);
-}
-//-------------------------------------------------------------
-void AttRelEnvelope::setRelease(float ms) {
-	releaseEnvelope_.setTc(ms);
-}
-//-------------------------------------------------------------
-void AttRelEnvelope::setSampleRate(float sampleRate) {
-	attackEnvelope_.setSampleRate(sampleRate);
-	releaseEnvelope_.setSampleRate(sampleRate);
-}
-
-//-------------------------------------------------------------
-// simple compressor
-//-------------------------------------------------------------
-SimpleComp::SimpleComp() : AttRelEnvelope(10.0, 100.0), threshdB_(0.0), ratio_(1.0), envdB_(DC_OFFSET) {
-}
-//-------------------------------------------------------------
-void SimpleComp::setThresh(float dB) {
-	threshdB_ = dB;
-}
-//-------------------------------------------------------------
-void SimpleComp::setRatio(float ratio) {
-	assert(ratio > 0.0);
-	ratio_ = ratio;
-}
-//-------------------------------------------------------------
-void SimpleComp::initRuntime(void) {
-	envdB_ = DC_OFFSET;
-}
-
-} // end namespace chunkware_simple
